refactor
Todo service 및 controller 로직 수정

+++ src/main/java/com/sparta/todo/domain/todo/service/TodoService.java
@@ -1,5 +1,7 @@
 package com.sparta.todo.domain.todo.service;
+
 import com.sparta.todo.api.WeatherService;
+import com.sparta.todo.domain.comment.repository.CommentRepository;
 import com.sparta.todo.domain.todo.dto.ModifyDto;
 import com.sparta.todo.domain.todo.dto.TodoRequestDto;
 import com.sparta.todo.domain.todo.dto.TodoResponseDto;
@@ -8,8 +10,8 @@ import com.sparta.todo.domain.todo.repository.TodoRepository;
 import com.sparta.todo.domain.user.entity.User;
 import com.sparta.todo.exception.CustomException;
 import com.sparta.todo.exception.ErrorCode;
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,49 +20,50 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class TodoService{
+public class TodoService {
 
     private final TodoRepository todoRepository;
     private final WeatherService weatherService;
+    private final CommentRepository commentRepository;
 
     public List<TodoResponseDto> todoFindAll(Pageable pageable) {
-        int offsetNum = pageable.getPageNumber();
-        int limitNum = pageable.getPageSize();
-        offsetNum *= 10;
-        return todoRepository.findAll(limitNum,offsetNum).stream().map(TodoResponseDto :: new).toList();
+            return todoRepository.findAll(pageable).map(todo ->{
+                todo.createCommentsCount(commentRepository.countByTodoId(todo.getId()));
+                return new TodoResponseDto(todo);
+            }).toList();
     }
 
-    @Transactional
-    public TodoResponseDto todoCreate(TodoRequestDto reqDto,HttpServletRequest request) {
-        User user = returnUser(request);
+    public TodoResponseDto todoCreate(TodoRequestDto reqDto, User user) {
         String weather = weatherService.getWeather();
-        Todo todo = Todo.from(reqDto.getTitle(),reqDto.getContent(),user,weather);
+        Todo todo = Todo.createTodo(reqDto.getTitle(), reqDto.getContent(), user, weather);
 
         return new TodoResponseDto(todoRepository.save(todo));
     }
 
     @Transactional
-    public void todoModify(Long id, ModifyDto modifyDto) {
+    public void todoModify(Long id, ModifyDto modifyDto, User user) {
         Todo findTodo = isValidId(id);
-        findTodo.modify(modifyDto.getTitle(),modifyDto.getContent());
+        if (findTodo.isValidWriteUser(user.getId())) {
+            findTodo.modify(modifyDto.getTitle(), modifyDto.getContent());
+        }
     }
 
-    @Transactional
-    public void todoDelete(Long id) {
-        isValidId(id);
-        todoRepository.deleteById(id);
+    public void todoDelete(Long id, User user) {
+        Todo findTodo = isValidId(id);
+        if (findTodo.isValidWriteUser(user.getId())) {
+            todoRepository.deleteById(id);
+        }
     }
 
-    public Todo todoOne(Long id) {
-        return isValidId(id);
+    public Todo todoFindById(Long id) {
+        Todo todo = isValidId(id);
+        todo.createCommentsCount(commentRepository.countByTodoId(id));
+        return todo;
     }
 
     private Todo isValidId(Long id) {
-       return todoRepository.findById(id).orElseThrow(()-> new CustomException(ErrorCode.NOT_TODO_ID));
-    }
-
-    private User returnUser(HttpServletRequest request) {
-        return (User)request.getAttribute("user");
+        return todoRepository.findById(id)
+                             .orElseThrow(() -> new CustomException(ErrorCode.NOT_TODO_ID));
     }
 }
